Fix TM test StatsMarshall to move to threadsafe
